+++ include/uxr/agent/transport/discovery/DiscoveryServerLinux.hpp
@@ -19,7 +19,7 @@
 
 #include <thread>
 #include <atomic>
-#include <poll.h>
+#include <sys/poll.h>
 
 namespace eprosima {
 namespace uxr {
