+++ include/uxr/agent/transport/tcp/TCPEndPoint.hpp
@@ -12,21 +12,41 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef _UXR_AGENT_TRANSPORT_UDP_ENDPOINT_HPP_
-#define _UXR_AGENT_TRANSPORT_UDP_ENDPOINT_HPP_
+#ifndef _UXR_AGENT_TRANSPORT_TCP_ENDPOINT_HPP_
+#define _UXR_AGENT_TRANSPORT_TCP_ENDPOINT_HPP_
 
 #include <uxr/agent/transport/EndPoint.hpp>
 
 #include <stdint.h>
+#include <vector>
+#include <mutex>
 
 namespace eprosima {
 namespace uxr {
 
-class UDPEndPoint : public EndPoint
+typedef enum TCPInputBufferState
+{
+    TCP_BUFFER_EMPTY,
+    TCP_SIZE_INCOMPLETE,
+    TCP_SIZE_READ,
+    TCP_MESSAGE_INCOMPLETE,
+    TCP_MESSAGE_AVAILABLE
+
+} TCPInputBufferState;
+
+struct TCPInputBuffer
+{
+    std::vector<uint8_t> buffer;
+    uint16_t position;
+    TCPInputBufferState state;
+    uint16_t msg_size;
+};
+
+class TCPEndPoint : public EndPoint
 {
 public:
-    UDPEndPoint(uint32_t addr, uint16_t port) : addr_(addr), port_(port) {}
-    ~UDPEndPoint() = default;
+    TCPEndPoint(uint32_t addr, uint16_t port) : addr_(addr), port_(port) {}
+    ~TCPEndPoint() = default;
 
     uint32_t get_addr() const { return addr_; }
     uint16_t get_port() const { return port_; }
@@ -39,4 +59,4 @@ private:
 } // namespace uxr
 } // namespace eprosima
 
-#endif //_UXR_AGENT_TRANSPORT_UDP_ENDPOINT_HPP_
+#endif //_UXR_AGENT_TRANSPORT_TCP_ENDPOINT_HPP_
