+++ include/uxr/agent/transport/tcp/TCPServerWindows.hpp
@@ -15,7 +15,7 @@
 #ifndef _UXR_AGENT_TRANSPORT_TCP_SERVER_HPP_
 #define _UXR_AGENT_TRANSPORT_TCP_SERVER_HPP_
 
-#include <uxr/agent/transport/Server.hpp>
+#include <uxr/agent/transport/tcp/TCPServer.hpp>
 #include <uxr/agent/config.hpp>
 #include <unordered_map>
 #include <winsock2.h>
@@ -26,27 +26,6 @@
 
 namespace eprosima {
 namespace uxr {
-
-/**************************************************************************************************
- * TCP EndPoint.
- **************************************************************************************************/
-typedef enum TCPInputBufferState
-{
-    TCP_BUFFER_EMPTY,
-    TCP_SIZE_INCOMPLETE,
-    TCP_SIZE_READ,
-    TCP_MESSAGE_INCOMPLETE,
-    TCP_MESSAGE_AVAILABLE
-
-} TCPInputBufferState;
-
-struct TCPInputBuffer
-{
-    std::vector<uint8_t> buffer;
-    uint16_t position;
-    TCPInputBufferState state;
-    uint16_t msg_size;
-};
 
 struct TCPConnection
 {
@@ -59,33 +38,11 @@ struct TCPConnection
     std::mutex mtx;
 };
 
-class TCPEndPoint : public EndPoint
+class TCPServerBase : public TCPServerBase
 {
 public:
-    TCPEndPoint(uint32_t addr, uint16_t port) : addr_(addr), port_(port) {}
-    ~TCPEndPoint() = default;
-
-    uint32_t get_addr() const { return addr_; }
-    uint16_t get_port() const { return port_; }
-
-private:
-    uint32_t addr_;
-    uint16_t port_;
-};
-
-/**************************************************************************************************
- * TCP Server.
- **************************************************************************************************/
-class TCPServer : public Server
-{
-public:
-    microxrcedds_agent_DllAPI TCPServer(uint16_t port);
-    microxrcedds_agent_DllAPI ~TCPServer() = default;
-
-    virtual void on_create_client(EndPoint* source, const dds::xrce::ClientKey& client_key) override;
-    virtual void on_delete_client(EndPoint* source) override;
-    virtual const dds::xrce::ClientKey get_client_key(EndPoint *source) override;
-    virtual std::unique_ptr<EndPoint> get_source(const dds::xrce::ClientKey& client_key) override;
+    microxrcedds_agent_DllAPI TCPServerBase(uint16_t port);
+    microxrcedds_agent_DllAPI ~TCPServerBase() = default;
 
 private:
     virtual bool init() override;
@@ -104,7 +61,6 @@ private:
     static int send_locking(TCPConnection& connection, char* buffer, int len);
 
 private:
-    uint16_t port_;
     std::array<TCPConnection, TCP_MAX_CONNECTIONS> connections_;
     std::set<uint32_t> active_connections_;
     std::list<uint32_t> free_connections_;
@@ -112,10 +68,6 @@ private:
     struct pollfd listener_poll_;
     std::array<struct pollfd, TCP_MAX_CONNECTIONS> poll_fds_;
     uint8_t buffer_[TCP_TRANSPORT_MTU];
-    std::unordered_map<uint64_t, uint32_t> source_to_connection_map_;
-    std::unordered_map<uint64_t, uint32_t> source_to_client_map_;
-    std::unordered_map<uint32_t, uint64_t> client_to_source_map_;
-    std::mutex clients_mtx_;
     std::unique_ptr<std::thread> listener_thread_;
     std::atomic<bool> running_cond_;
     std::queue<InputPacket> messages_queue_;
